requirements.txt
openpyxl==3.1.5
pandas==2.2.3
pdf2image==1.17.0
PyPDF2==3.0.1
python-dotenv==1.1.0
Requests==2.32.3
streamlit==1.45.0
